Added a helper `Assert` method - `assertEqualHStore` for comparing HStore fields.
The reason we need one is because when you query a object from the DB, his HStore's values do not get converted into the respectable types. e.g { "hello": "2"}, when 2 needs to be an integer. Same thing for nested dicts - {"hello": "{"yo": "yo"}"}, when the inner `yo` dict needed to be of type `dict`